Moved StaticInfo.java and JumpInfo to the common package
